+++ pyunlocbox/demos/__init__.py
@@ -1,0 +1,2 @@
+from pyunlocbox.demos import demo_douglas_rachford
+assert demo_douglas_rachford