+++ HISTORY.rst
@@ -3,6 +3,17 @@
 =======
 History
 =======
+0.2.2 (2015-01-16)
+------------------
+
+New feature version. Still experimental.
+
+New Features:
+
+* norm_tv has been added with gradient, div, evaluation and prox.
+* Module signals has been added.
+* A demo for douglas rachford is also now presenta.
+
 
 0.2.1 (2014-08-20)
 ------------------
@@ -43,6 +54,7 @@ Features :
 * Forward-backward splitting algorithm
 * L1-norm function (eval and prox)
 * L2-norm function (eval, grad and prox)
+* TV-norm function (eval, grad, div and prox)
 * Least square problem tutorial using L2-norm and forward-backward
 * Compressed sensing tutorial using L1-norm, L2-norm and forward-backward